+++ NEWS.rst
@@ -6,7 +6,9 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Major: Removed offset from sample storage. This functionality must now be
+  performed outside of the SampleStorage.
+
 
 1.0.0
 -----